Split AuthCoordinator under the file length budget
The offline sign-out work grew AuthCoordinator.swift to 1111 lines
against its tracked 757 budget. The Priming section (prime, existing
session check, auto-login, cached-session validation, stale clears)
moves to AuthCoordinator+SessionPriming.swift and the Tokens section to
AuthCoordinator+Tokens.swift. Members the split files share drop
`private` for module-internal access (CmuxAuthRuntime is its own
module, so this stays package-private in effect); public setters become
internal(set). No behavior changes.
